+++ crates/reinhardt-commands/CHANGELOG.md
@@ -27,6 +27,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - N/A
 
 
+## [0.1.0-alpha.2] - 2026-01-28
+
+### Changed
+- Migrated welcome page rendering from Tera to reinhardt-pages SSR
+- Added reinhardt-pages dependency
+
+### Removed
+- Removed welcome.tpl template (replaced by WelcomePage component)
+
+
 ## [0.1.0-alpha.1] - 2026-01-23
 
 ### Added